move griberies paths functions to a  sub-module
